requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
duckdb==0.10.2
pandas==2.2.2
polars==0.20.25
yfinance==0.2.40
python-dotenv==1.0.1
pyarrow==16.0.0
redis==5.0.1